Абстрактный скролл контроллер https://online.sbis.ru/opendoc.html?guid=1aa301a5-bca0-4e59-90f2-71ce15c12dba
