+++ Controls/_baseList/Controllers/ScrollController/ObserverController/ObserversController.ts
@@ -1,7 +1,9 @@
-import {AbstractObserversController, TIntersectionEvent} from './AbstractObserversController';
+import {AbstractObserversController, IAbstractObserversControllerOptions, TIntersectionEvent} from './AbstractObserversController';
 import {EdgeIntersectionObserver} from 'Controls/scroll';
 import {Control} from 'UI/Base';
 import {IDirection} from './../ScrollController';
+
+export {IAbstractObserversControllerOptions as IObserversControllerOptions};
 
 export class ObserversController extends AbstractObserversController {
     protected _createTriggersObserver(component: Control,
