+++ Controls/_baseList/Controllers/ScrollController/ScrollController.ts
@@ -1,6 +1,6 @@
 import {
     IItemsSizes,
-    IAbstarctItemsSizesControllerOptions,
+    IAbstractItemsSizesControllerOptions,
     AbstractItemsSizesController
 } from './ItemsSizeController/AbstractItemsSizeController';
 import {
@@ -101,11 +101,11 @@ export type IHasItemsOutRangeChangedCallback = (hasItems: IHasItemsOutRange) => 
 export type IPlaceholdersChangedCallback = (placeholders: IPlaceholders) => void;
 
 export interface IScrollControllerOptions extends
-    IAbstarctItemsSizesControllerOptions,
+    IAbstractItemsSizesControllerOptions,
     IAbstractObserversControllerBaseOptions,
     ICalculatorBaseOptions {
     observerControllerConstructor: new (options: IAbstractObserversControllerOptions) => AbstractObserversController;
-    itemsSizeControllerConstructor: new (options: IAbstarctItemsSizesControllerOptions) => AbstractItemsSizesController;
+    itemsSizeControllerConstructor: new (options: IAbstractItemsSizesControllerOptions) => AbstractItemsSizesController;
     indexesInitializedCallback: IIndexesInitializedCallback;
     indexesChangedCallback: IIndexesChangedCallback;
     activeElementChangedCallback: IActiveElementChangedChangedCallback;
