+++ Controls/baseList.ts
@@ -88,6 +88,11 @@ export {MoveController, IMoveControllerOptions}  from 'Controls/_baseList/Contro
 
 //region new scroll
 export {
+    AbstractListVirtualScrollController,
+    IAbstractListVirtualScrollControllerOptions
+} from './_baseList/Controllers/AbstractListVirtualScrollController';
+
+export {
     ScrollController as NewScrollController,
     IDirection as IDirectionNew,
     IPlaceholders
@@ -95,7 +100,7 @@ export {
 export {
     IItemSize,
     AbstractItemsSizesController,
-    IAbstarctItemsSizesControllerOptions
+    IAbstractItemsSizesControllerOptions
 } from './_baseList/Controllers/ScrollController/ItemsSizeController/AbstractItemsSizeController';
 export {
     AbstractObserversController,
